Finalize GL API split and migrate legacy functions
