+++ src/gl_api_pixels.h
@@ -9,5 +9,10 @@ GL_API void GL_APIENTRY glReadPixels(GLint x, GLint y, GLsizei width,
 GL_API void GL_APIENTRY glColorMask(GLboolean r, GLboolean g, GLboolean b,
 				    GLboolean a);
 GL_API void GL_APIENTRY glDepthRangef(GLfloat n, GLfloat f);
+GL_API void GL_APIENTRY glClearColor(GLfloat r, GLfloat g, GLfloat b,
+				     GLfloat a);
+GL_API void GL_APIENTRY glClearColorx(GLfixed r, GLfixed g, GLfixed b,
+				      GLfixed a);
+GL_API void GL_APIENTRY glPointSize(GLfloat size);
 
 #endif /* GL_API_PIXELS_H */
