+++ src/gl_memory_tracker.c
@@ -5,18 +5,19 @@
 #include <stdbool.h>
 #include <threads.h>
 
-#define MAX_ALLOCATIONS 1000
+#define INITIAL_ALLOCATIONS 1024
 
 typedef struct {
 	void *ptr;
 	size_t size;
 	stage_tag_t stage;
-	const char *type;
+	char *type;
 	int line;
 	bool active;
 } allocation_t;
 
-static allocation_t g_allocs[MAX_ALLOCATIONS];
+static allocation_t *g_allocs;
+static size_t g_capacity;
 static mtx_t g_mutex;
 static size_t g_current;
 static size_t g_peak;
@@ -26,7 +27,10 @@ static size_t g_stage_peak[STAGE_COUNT];
 int memory_tracker_init(void)
 {
 	mtx_init(&g_mutex, mtx_plain);
-	memset(g_allocs, 0, sizeof(g_allocs));
+	g_capacity = INITIAL_ALLOCATIONS;
+	g_allocs = calloc(g_capacity, sizeof(allocation_t));
+	if (!g_allocs)
+		return 0;
 	g_current = g_peak = 0;
 	memset(g_stage_current, 0, sizeof(g_stage_current));
 	memset(g_stage_peak, 0, sizeof(g_stage_peak));
@@ -37,10 +41,11 @@ int memory_tracker_init(void)
 static void record_alloc(void *ptr, size_t size, stage_tag_t stage,
 			 const char *type, int line)
 {
-	for (int i = 0; i < MAX_ALLOCATIONS; ++i) {
+	for (size_t i = 0; i < g_capacity; ++i) {
 		if (!g_allocs[i].active) {
-			g_allocs[i] = (allocation_t){ ptr,  size, stage,
-						      type, line, true };
+			g_allocs[i] =
+				(allocation_t){ ptr,	      size, stage,
+						strdup(type), line, true };
 			g_current += size;
 			g_stage_current[stage] += size;
 			if (g_current > g_peak)
@@ -50,7 +55,26 @@ static void record_alloc(void *ptr, size_t size, stage_tag_t stage,
 			return;
 		}
 	}
-	LogMessage(LOG_LEVEL_ERROR, "Memory tracker pool exhausted");
+	size_t new_cap = g_capacity * 2;
+	allocation_t *new_allocs =
+		realloc(g_allocs, new_cap * sizeof(allocation_t));
+	if (!new_allocs) {
+		LogMessage(LOG_LEVEL_ERROR, "Memory tracker pool exhausted");
+		return;
+	}
+	memset(new_allocs + g_capacity, 0,
+	       (new_cap - g_capacity) * sizeof(allocation_t));
+	g_allocs = new_allocs;
+	size_t idx = g_capacity;
+	g_capacity = new_cap;
+	g_allocs[idx] =
+		(allocation_t){ ptr, size, stage, strdup(type), line, true };
+	g_current += size;
+	g_stage_current[stage] += size;
+	if (g_current > g_peak)
+		g_peak = g_current;
+	if (g_stage_current[stage] > g_stage_peak[stage])
+		g_stage_peak[stage] = g_stage_current[stage];
 }
 
 void *memory_tracker_allocate(size_t size, stage_tag_t stage, const char *type,
@@ -81,23 +105,38 @@ void *memory_tracker_realloc(void *ptr, size_t size, stage_tag_t stage,
 			     const char *type, int line)
 {
 	mtx_lock(&g_mutex);
+	allocation_t *rec = NULL;
 	if (ptr) {
-		for (int i = 0; i < MAX_ALLOCATIONS; ++i) {
+		for (size_t i = 0; i < g_capacity; ++i) {
 			if (g_allocs[i].active && g_allocs[i].ptr == ptr) {
-				g_current -= g_allocs[i].size;
-				g_stage_current[g_allocs[i].stage] -=
-					g_allocs[i].size;
-				g_allocs[i].active = false;
+				rec = &g_allocs[i];
 				break;
 			}
 		}
 	}
-	mtx_unlock(&g_mutex);
 	void *np = realloc(ptr, size);
-	if (!np)
+	if (!np) {
+		mtx_unlock(&g_mutex);
 		return NULL;
-	mtx_lock(&g_mutex);
-	record_alloc(np, size, stage, type, line);
+	}
+	if (rec) {
+		g_current -= rec->size;
+		g_stage_current[rec->stage] -= rec->size;
+		free(rec->type);
+		*rec =
+			(allocation_t){ np,	      size, stage,
+					strdup(type), line, true };
+	} else {
+		record_alloc(np, size, stage, type, line);
+		mtx_unlock(&g_mutex);
+		return np;
+	}
+	g_current += size;
+	g_stage_current[stage] += size;
+	if (g_current > g_peak)
+		g_peak = g_current;
+	if (g_stage_current[stage] > g_stage_peak[stage])
+		g_stage_peak[stage] = g_stage_current[stage];
 	mtx_unlock(&g_mutex);
 	return np;
 }
@@ -111,11 +150,12 @@ void memory_tracker_free(void *ptr, stage_tag_t stage, const char *type,
 	if (!ptr)
 		return;
 	mtx_lock(&g_mutex);
-	for (int i = 0; i < MAX_ALLOCATIONS; ++i) {
+	for (size_t i = 0; i < g_capacity; ++i) {
 		if (g_allocs[i].active && g_allocs[i].ptr == ptr) {
 			g_current -= g_allocs[i].size;
 			g_stage_current[g_allocs[i].stage] -= g_allocs[i].size;
 			g_allocs[i].active = false;
+			free(g_allocs[i].type);
 			break;
 		}
 	}
@@ -146,12 +186,13 @@ void memory_tracker_report(void)
 void memory_tracker_shutdown(void)
 {
 	mtx_lock(&g_mutex);
-	for (int i = 0; i < MAX_ALLOCATIONS; ++i) {
+	for (size_t i = 0; i < g_capacity; ++i) {
 		if (g_allocs[i].active) {
 			LogMessage(LOG_LEVEL_WARN,
 				   "Leaked %zu bytes from %s:%d",
 				   g_allocs[i].size, g_allocs[i].type,
 				   g_allocs[i].line);
+			free(g_allocs[i].type);
 		}
 	}
 	for (int s = 0; s < STAGE_COUNT; ++s) {
@@ -160,5 +201,6 @@ void memory_tracker_shutdown(void)
 	}
 	LogMessage(LOG_LEVEL_INFO, "Peak memory usage %zu bytes", g_peak);
 	mtx_unlock(&g_mutex);
+	free(g_allocs);
 	mtx_destroy(&g_mutex);
 }
