+++ src/pipeline/gl_primitive.c
@@ -1,6 +1,8 @@
 #include "gl_primitive.h"
 #include "gl_raster.h"
 #include "../gl_thread.h"
+#define PIPELINE_USE_GLSTATE 0
+_Static_assert(PIPELINE_USE_GLSTATE == 0, "pipeline must not touch gl_state");
 #include "../gl_memory_tracker.h"
 
 void pipeline_assemble_triangle(Triangle *dst, const Vertex *v0,
