+++ src/pipeline/gl_raster.c
@@ -1,6 +1,8 @@
 #include "gl_raster.h"
 #include "gl_fragment.h"
 #include "../gl_logger.h"
+#define PIPELINE_USE_GLSTATE 0
+_Static_assert(PIPELINE_USE_GLSTATE == 0, "pipeline must not touch gl_state");
 #include "../gl_memory_tracker.h"
 #include "../gl_thread.h"
